refactor: extract UI/utility functions into lib/ui-utils.ts (Phase 5)
Extract 24 UI/utility functions from content.ts into a dedicated
lib/ui-utils.ts module using the DI factory pattern established in
Phases 3a-4. Functions include export/metadata helpers, auto-sync UI
functions, collection helpers, and general utilities.

content.ts: 3,277L → 2,986L (-291 lines net after factory wiring)
make check: PASS
